test(auth): 인증 관련 테스트 코드 한 파일에 모아놓도록 수정
- 테스트 실행 시간 단축 및 보일러 코드 제거를 위해 한 파일에 테스트 코드 모아놓도록 수정
    - 테스트 모듈 세팅, 앱 및 db 초기화 등의 코드가 중복해서 들어가야 하고, 테스트 파일을 여러 개로 쪼개 놓으면 실행 세팅이 오래 걸림
